dubai_real_estate_workflow/tools/scraping_tools.py
"""
Web scraping tools for extracting leads from various real estate platforms
"""

import os
import time
import logging
from typing import List, Dict, Any
from crewai_tools import BaseTool
import requests
from bs4 import BeautifulSoup
import json

logger = logging.getLogger(__name__)


class LinkedInScraperTool(BaseTool):
    name: str = "LinkedIn Lead Scraper"
    description: str = (
        "Scrapes LinkedIn for property buyer leads and mortgage inquiries in Dubai. "
        "Extracts contact name, profile URL, and interest indicators."
    )
    
    def _run(self, search_keywords: List[str], max_results: int = 50) -> List[Dict[str, Any]]:
        """
        Scrape LinkedIn for leads based on search keywords
        
        Args:
            search_keywords: List of keywords to search for
            max_results: Maximum number of results to return
            
        Returns:
            List of lead dictionaries with contact information
        """
        leads = []
        
        # Note: This is a simplified implementation
        # In production, use LinkedIn API or proper authentication
        logger.info(f"Scraping LinkedIn with keywords: {search_keywords}")
        
        try:
            # Simulated scraping logic (replace with actual implementation)
            for keyword in search_keywords:
                # In production, implement proper LinkedIn scraping with authentication
                # For now, returning mock data structure
                mock_leads = self._mock_linkedin_data(keyword, max_results // len(search_keywords))
                leads.extend(mock_leads)
                
                time.sleep(2)  # Rate limiting
                
            logger.info(f"Successfully scraped {len(leads)} leads from LinkedIn")
            return leads
            
        except Exception as e:
            logger.error(f"Error scraping LinkedIn: {str(e)}")
            return []
    
    def _mock_linkedin_data(self, keyword: str, count: int) -> List[Dict[str, Any]]:
        """Generate mock LinkedIn data for demonstration"""
        return [
            {
                "source": "LinkedIn",
                "name": f"Lead {i}",
                "profile_url": f"https://linkedin.com/in/lead-{i}",
                "search_keyword": keyword,
                "interest_level": "high" if i % 3 == 0 else "medium",
                "timestamp": time.time()
            }
            for i in range(count)
        ]


class PropertyFinderScraperTool(BaseTool):
    name: str = "Property Finder Scraper"
    description: str = (
        "Scrapes Property Finder for buyer inquiries and contact information. "
        "Extracts name, email, phone, budget, and preferred areas."
    )
    
    def _run(self, location: str = "dubai", category: str = "residential-for-sale") -> List[Dict[str, Any]]:
        """
        Scrape Property Finder for leads
        
        Args:
            location: Target location (default: dubai)
            category: Property category
            
        Returns:
            List of lead dictionaries
        """
        leads = []
        base_url = "https://www.propertyfinder.ae"
        
        logger.info(f"Scraping Property Finder: {location}/{category}")
        
        try:
            # In production, implement actual scraping with proper headers and authentication
            # This is a simplified mock implementation
            
            headers = {
                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
            }
            
            # Mock implementation - replace with actual scraping
            leads = self._mock_property_finder_data()
            
            logger.info(f"Successfully scraped {len(leads)} leads from Property Finder")
            return leads
            
        except Exception as e:
            logger.error(f"Error scraping Property Finder: {str(e)}")
            return []
    
    def _mock_property_finder_data(self) -> List[Dict[str, Any]]:
        """Generate mock Property Finder data"""
        return [
            {
                "source": "Property Finder",
                "name": f"Buyer {i}",
                "email": f"buyer{i}@example.com",
                "phone": f"+971-50-{1000000 + i}",
                "budget": 800000 + (i * 100000),
                "preferred_area": ["Downtown Dubai", "Business Bay"][i % 2],
                "property_type": "apartment",
                "bedrooms": 1 + (i % 2),
                "timestamp": time.time()
            }
            for i in range(10)
        ]


class BayutScraperTool(BaseTool):
    name: str = "Bayut Scraper"
    description: str = (
        "Scrapes Bayut.com for property buyer leads in Dubai. "
        "Extracts contact details, budget range, and property preferences."
    )
    
    def _run(self, location: str = "dubai", purpose: str = "for-sale") -> List[Dict[str, Any]]:
        """
        Scrape Bayut for leads
        
        Args:
            location: Target location
            purpose: Property purpose (for-sale, for-rent)
            
        Returns:
            List of lead dictionaries
        """
        leads = []
        base_url = "https://www.bayut.com"
        
        logger.info(f"Scraping Bayut: {location}/{purpose}")
        
        try:
            # Mock implementation - replace with actual scraping
            leads = self._mock_bayut_data()
            
            logger.info(f"Successfully scraped {len(leads)} leads from Bayut")
            return leads
            
        except Exception as e:
            logger.error(f"Error scraping Bayut: {str(e)}")
            return []
    
    def _mock_bayut_data(self) -> List[Dict[str, Any]]:
        """Generate mock Bayut data"""
        areas = ["Azizi Riviera", "Dubai Marina", "JBR", "Business Bay"]
        return [
            {
                "source": "Bayut",
                "name": f"Prospect {i}",
                "email": f"prospect{i}@example.com",
                "phone": f"+971-55-{2000000 + i}",
                "budget": 700000 + (i * 150000),
                "preferred_area": areas[i % len(areas)],
                "property_type": "apartment",
                "bedrooms": 1 + (i % 2),
                "property_status": "off-plan" if i % 2 == 0 else "ready",
                "timestamp": time.time()
            }
            for i in range(10)
        ]


class DubizzleScraperTool(BaseTool):
    name: str = "Dubizzle Scraper"
    description: str = (
        "Scrapes Dubizzle for property buyer inquiries in Dubai. "
        "Extracts contact information and property requirements."
    )
    
    def _run(self, category: str = "apartments-flats") -> List[Dict[str, Any]]:
        """
        Scrape Dubizzle for leads
        
        Args:
            category: Property category
            
        Returns:
            List of lead dictionaries
        """
        leads = []
        base_url = "https://dubai.dubizzle.com/property-for-sale"
        
        logger.info(f"Scraping Dubizzle: {category}")
        
        try:
            # Mock implementation - replace with actual scraping
            leads = self._mock_dubizzle_data()
            
            logger.info(f"Successfully scraped {len(leads)} leads from Dubizzle")
            return leads
            
        except Exception as e:
            logger.error(f"Error scraping Dubizzle: {str(e)}")
            return []
    
    def _mock_dubizzle_data(self) -> List[Dict[str, Any]]:
        """Generate mock Dubizzle data"""
        areas = ["Palm Jumeirah", "Dubai Creek Harbour", "Dubai Hills Estate"]
        return [
            {
                "source": "Dubizzle",
                "name": f"Inquirer {i}",
                "email": f"inquirer{i}@example.com",
                "phone": f"+971-56-{3000000 + i}",
                "budget": 900000 + (i * 200000),
                "preferred_area": areas[i % len(areas)],
                "property_type": "apartment",
                "bedrooms": 1 + (i % 2),
                "timestamp": time.time()
            }
            for i in range(10)
        ]
